code/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
elastic-transport==8.4.0
elasticsearch==8.9.0
lxml==4.9.3
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==1.26.16
xmltodict==0.13.0
